acolyte/requirements.txt
Flask
Flask-SQLAlchemy
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
gunicorn
itsdangerous
Flask-WTF
beautifulsoup4
pytest
pyyaml
